requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
matplotlib
scikit-learn # For label_binarize, etc., if you want to include evaluation utilities later, though not strictly needed for just prediction
Pillow # torchvision.transforms.ToPILImage() uses this
streamlit>=1.28.0
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0